pizza_generator/pyproject.toml
[project]
name = "pizza-gen"
version = "0.1.0"
description = "pizza-clock/pizza_generator"
readme = "README.md"
authors = [
    { name = "likeablob", email = "46628917+likeablob@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.7",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "webuiapi>=0.9.17",
]

[dependency-groups]
dev = [
    "ruff>=0.9.5",
]

[project.scripts]
pizza_gen = "pizza_gen.cli:app"
circular_prompt_gen = "circular_prompt_gen.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pizza_gen", "src/circular_prompt_gen"]

[tool.uv]
package = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["I"]

